Update round2_final.py
I added a "correction" of your round 2 function at the bottom (in quotes so it doesn't run). Just a general outline of what it should look like, it will have to be adjusted to fit with the Deck code and stuff, but this should give you an idea of how the code should function I think.